author + move DiagramUtils
