+++ src/main/java/com/powsybl/sld/utils/DiagramUtils.java
@@ -1,4 +1,10 @@
-package com.powsybl.sld.server;
+/**
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.sld.utils;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.Network;
@@ -9,6 +15,9 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.UUID;
 
+/**
+ * @author Etienne Homer<etienne.homer at rte-france.com>
+ */
 public final class DiagramUtils {
     private DiagramUtils() {
         throw new AssertionError("Utility class should not be instantiated");
